requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.5.0

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3

# AI and ML
sentence-transformers==2.2.2
chromadb==0.4.18
ollama==0.1.7
crewai==0.11.0

# Free Model Alternatives
groq==0.4.2
transformers==4.36.0
torch==2.1.1
accelerate==0.25.0

# Data processing
pandas==2.1.4
numpy==1.25.2

# Utilities
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0 